+++ packages/media/src/index.ts
@@ -1,12 +1,8 @@
 import { Accessor } from "solid-js";
 import { isServer } from "solid-js/web";
 import { makeEventListener } from "@solid-primitives/event-listener";
-import {
-  createHydratableStaticStore,
-  entries,
-  noop,
-  createHydratableSignal,
-} from "@solid-primitives/utils";
+import { entries, noop, createHydratableSignal } from "@solid-primitives/utils";
+import { createHydratableStaticStore } from "@solid-primitives/static-store";
 import { createHydratableSingletonRoot } from "@solid-primitives/rootless";
 
 /**