+++ packages/static-store/dev/vite.config.ts
@@ -1,0 +1,2 @@
+import { viteConfig } from "../../../configs/vite.config";
+export default viteConfig;