+++ packages/autofocus/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @solid-primitives/autofocus
 
+## 0.0.105
+
+### Patch Changes
+
+- e0d87bcf: Make `autofocus` usable with `ref`
+
+## 0.0.104
+
+### Patch Changes
+
+- Updated dependencies [2f6d3732]
+  - @solid-primitives/utils@6.0.0
+
 ## 0.0.103
 
 ### Patch Changes