+++ packages/bounds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @solid-primitives/bounds
 
+## 0.0.111
+
+### Patch Changes
+
+- 2f6d3732: Use the static-store package for creating reactive objects. Minor overall improvements
+- Updated dependencies [2f6d3732]
+- Updated dependencies [2f6d3732]
+  - @solid-primitives/resize-observer@2.0.14
+  - @solid-primitives/utils@6.0.0
+  - @solid-primitives/event-listener@2.2.10
+  - @solid-primitives/static-store@0.0.2
+
 ## 0.0.110
 
 ### Patch Changes