+++ packages/intersection-observer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/intersection-observer
 
+## 2.0.11
+
+### Patch Changes
+
+- Updated dependencies [2f6d3732]
+  - @solid-primitives/utils@6.0.0
+
 ## 2.0.10
 
 ### Patch Changes