+++ packages/refs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/refs
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [2f6d3732]
+  - @solid-primitives/utils@6.0.0
+
 ## 1.0.1
 
 ### Patch Changes